crop-recommendation-system/requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==1.7.6
matplotlib==3.7.2
seaborn==0.12.2
streamlit==1.25.0
plotly==5.15.0
shap==0.42.1
requests==2.31.0
python-dotenv==1.0.0
joblib==1.3.1
Pillow==10.0.0
openpyxl==3.1.2
